+++ charmcraft/commands/build.py
@@ -27,7 +27,7 @@ import yaml
 
 from charmcraft.cmdbase import BaseCommand, CommandError
 from charmcraft.jujuignore import JujuIgnore, default_juju_ignore
-from .utils import make_executable
+from charmcraft.utils import make_executable
 
 logger = logging.getLogger(__name__)
